fixed stupid save bug

+++ bitpix-move/ViewController.m
@@ -113,7 +113,7 @@ static BOOL _isPreviewing = NO;
     
     NSDictionary *animationInfo = [NSDictionary dictionaryWithObjects:[NSArray arrayWithObjects:self.uuid, today, frames, nil] forKeys:[NSArray arrayWithObjects:@"name", @"date", @"frames", nil]];
     
-    int index = [self.appData.userAnimations indexOfObjectPassingTest:^BOOL(id  _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
+    NSUInteger index = [self.appData.userAnimations indexOfObjectPassingTest:^BOOL(id  _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
         NSDictionary *animation = (NSDictionary *)obj;
         BOOL found = [[animation objectForKey:@"name"] isEqualToString:self.uuid];
         return found;
